# backend/tests/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ApiKeyCreate": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "key"
        ],
        "title": "ApiKeyCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationConfigCreate": {
        "properties": {
          "action_tier": {
            "default": 0,
            "title": "Action Tier",
            "type": "integer"
          },
          "adapter_type": {
            "title": "Adapter Type",
            "type": "string"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "provider_name": {
            "title": "Provider Name",
            "type": "string"
          }
        },
        "required": [
          "adapter_type",
          "provider_name"
        ],
        "title": "IntegrationConfigCreate",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings/api-keys/{provider}": {
      "delete": {
        "operationId": "delete_api_key_api_settings_api_keys__provider__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Api Key",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "set_api_key_api_settings_api_keys__provider__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Api Key",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/integrations": {
      "get": {
        "operationId": "list_integrations_endpoint_api_settings_integrations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationConfig"
                  },
                  "title": "Response List Integrations Endpoint Api Settings Integrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Integrations Endpoint",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_integration_endpoint_api_settings_integrations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Integration Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/settings/integrations/registry/{entry_id}": {
      "get": {
        "description": "Get a single registry entry with full setup guide.",
        "operationId": "get_registry_entry_endpoint_api_settings_integrations_registry__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Registry Entry Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/integrations/{integration_id}": {
      "delete": {
        "operationId": "delete_integration_endpoint_api_settings_integrations__integration_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "title": "Integration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Integration Endpoint",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_integration_endpoint_api_settings_integrations__integration_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "title": "Integration Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Integration Endpoint",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/integrations/{integration_id}/credentials": {
      "get": {
        "description": "List credential key names for an integration (no values).",
        "operationId": "list_credentials_api_settings_integrations__integration_id__credentials_get",
        "parameters": [
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "title": "Integration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialInfo"
                  },
                  "title": "Response List Credentials Api Settings Integrations  Integration Id  Credentials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Credentials",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "description": "Store an encrypted credential for an integration.",
        "operationId": "store_credential_api_settings_integrations__integration_id__credentials_post",
        "parameters": [
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "title": "Integration Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Store Credential",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/integrations/{integration_id}/credentials/{key_name}": {
      "delete": {
        "description": "Delete a single credential.",
        "operationId": "delete_credential_api_settings_integrations__integration_id__credentials__key_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "title": "Integration Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_name",
            "required": true,
            "schema": {
              "title": "Key Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Credential",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}